requirements.txt
Flask==3.0.0
plotly==5.18.0
python-dotenv==1.0.0
requests==2.31.0
reportlab==4.0.7
matplotlib==3.8.2